service-workers/service-worker/registration-iframe.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Service Worker: Registration for iframe</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/test-helpers.sub.js"></script>
<script>

// Set script url and scope url relative to the iframe's document's url. Assert
// the implementation parses the urls against the iframe's document's url.
async_test(function(t) {
  var url = 'resources/blank.html';
  var scope = 'registration-for-iframe-from-parent-frame';
  var expected_scope = normalizeURL('resources/' + scope);
  var script = 'empty-worker.js';
  var expected_script = normalizeURL('resources/' + script);
  var frame;
  var registration;

  service_worker_unregister(t, scope)
    .then(function() { return with_iframe(url); })
    .then(function(f) {
        frame = f;
        return frame.contentWindow.navigator.serviceWorker.register(
            script,
            { scope: scope });
      })
    .then(function(r) {
        registration = r;
        return wait_for_state(t, r.installing, 'activated');
      })
    .then(function() {
        assert_equals(registration.scope, expected_scope,
                      'registration\'s scope must be parsed against the ' +
                      '"relevant global object"');
        assert_equals(registration.active.scriptURL, expected_script,
                      'worker\'s scriptURL must be parsed against the ' +
                      '"relevant global object"');
        frame.remove();
        return service_worker_unregister_and_done(t, scope);
      })
    .catch(unreached_rejection(t));
  }, 'register method should use the "relevant global object" to parse its ' +
     'scriptURL and scope - normal case');

// Set script url and scope url relative to the parent frame's document's url.
// Assert the implementation throws a TypeError exception.
async_test(function(t) {
  var url = 'resources/blank.html';
  var scope = 'resources/registration-for-iframe-from-parent-frame';
  var script = 'resources/empty-worker.js';
  var frame;
  var registration;

  service_worker_unregister(t, scope)
    .then(function() { return with_iframe(url); })
    .then(function(f) {
        frame = f;
        return frame.contentWindow.navigator.serviceWorker.register(
            script,
            { scope: scope });
      })
    .then(
      function() {
        assert_unreached('register() should reject');
      },
      function(e) {
        assert_equals(e.name, 'TypeError',
                      'register method with scriptURL and scope parsed to ' +
                      'nonexistent location should reject with TypeError');
        frame.remove();
        return service_worker_unregister_and_done(t, scope);
      })
    .catch(unreached_rejection(t));
  }, 'register method should use the "relevant global object" to parse its ' +
     'scriptURL and scope - error case');

// Set the scope url to a non-subdirectory of the script url. Assert the
// implementation throws a SecurityError exception.
async_test(function(t) {
  var url = 'resources/blank.html';
  var scope = '../registration-for-iframe-from-parent-frame';
  var script = 'empty-worker.js';
  var frame;
  var registration;

  service_worker_unregister(t, scope)
    .then(function() { return with_iframe(url); })
    .then(function(f) {
        frame = f;
        return frame.contentWindow.navigator.serviceWorker.register(
            script,
            { scope: scope });
      })
    .then(
      function() {
        assert_unreached('register() should reject');
      },
      function(e) {
        assert_equals(e.name, 'SecurityError',
                      'The scope set to a non-subdirectory of the scriptURL ' +
                      'should reject with SecurityError');
        frame.remove();
        return service_worker_unregister_and_done(t, scope);
      })
    .catch(unreached_rejection(t));
  }, 'A scope url should start with the given script url');

</script>
